+++ CHANGELOG.md
@@ -9,4 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Two separate FastAPI apps, webapp & apiapp, to keep the API separate from serving the frontend.
 - Minimal FastAPI application.